packages/backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api_zenith"
version = "0.1.0"
description = "Backend service for API Zenith"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "Flask",
    "python-dotenv",
    "psycopg2-binary",
    "redis"
]
